3/19 打卡记录TODO

+++ app/src/main/java/com/habitree/xueshu/mine/bean/Wallet.java
@@ -5,13 +5,13 @@ package com.habitree.xueshu.mine.bean;
 public class Wallet {
     /**
      * mem_id : 3
-     * balance : 466.50
+     * balance : 420.50
      * frozen_money : 0.00
      * sum_money : 0.00
      * last_charge_time : 0
      * last_pay_time : 0
-     * last_money : 0
-     * update_time : 2018-03-01 23:17:04
+     * last_money : 427.5
+     * update_time : 2018-03-19 18:06:38
      */
 
     public int mem_id;
@@ -20,6 +20,6 @@ public class Wallet {
     public String sum_money;
     public int last_charge_time;
     public int last_pay_time;
-    public int last_money;
+    public double last_money;
     public String update_time;
 }
